added header file and merge sort
Moved functions into a separate header file because main was getting a little crowded. Also implemented Merge Sort. Will implement another 1 or 2 sorting algos later.
